# app/blueprints/customers/routes.py
from .schemas import customer_schema, customers_schema, login_schema
from flask import request, jsonify
from marshmallow import ValidationError
from sqlalchemy import select
from app.models import Customer, db
from .import customers_bp
from app.extensions import limiter, cache
from app.utils.utils import encode_token, token_required

@customers_bp.route("/login", methods =['POST'])
def login():
     
     try:
          credentials = login_schema.load(request.json)
          email = credentials['email']
          password = credentials['password']
     except ValidationError as e:
          return jsonify(e.messages), 400
     
     query = select(Customer).where(Customer.email == email)
     customer = db.session.execute(query).scalars().first()

     if customer and customer.password == password:
          token = encode_token(customer.id)

          response = {
               "status": "success", 
               "message": "Successfully logged in.",
               "token": token
          }

          return jsonify(response), 200
     else: 
          return jsonify({"message": "Invalid email or password!"})


#Create a new customer

@customers_bp.route("/", methods=['POST'])
@limiter.limit("2 per hour") # Rate limiting: 3 requests per hour per IP  
def create_customer():
     try: 
          customer_data = customer_schema.load(request.json) # Validate and deserialize input
     except ValidationError as e:
          return jsonify(e.messages),400
     
     query = select(Customer). where(Customer.email == customer_data['email']) #checking our db for a customer with this email
     existing_customer = db.session.execute(query).scalars().all()
     if existing_customer:
          return jsonify({"error": "Email already associated with another account."}), 400
     
     new_customer = Customer(name=customer_data['name'], email=customer_data['email'],
                             phone=customer_data['phone'], password=customer_data['password'])
     db.session.add(new_customer)
     db.session.commit()
     return customer_schema.jsonify(new_customer), 201

#Get all customers
@customers_bp.route("/", methods=['GET'])
#@cache.cached(timeout =60)  # Cache the response for 60 seconds
def get_customers():
     try: 
        page = int(request.args.get('page'))
        per_page = int(request.args.get('per_page'))  
        query = select(Customer)
        customers = db.paginate(query, page=page, per_page=per_page)
        return customers_schema.jsonify(customers), 200
     except:     
          query = select(Customer)
          customers = db.session.execute(query).scalars().all()
          return customers_schema.jsonify(customers)

#Get specific customer 
@customers_bp.route("/<int:customer_id>", methods=['GET'])
def get_customer(customer_id):
     customer = db.session.get(Customer, customer_id)

     if customer:
          return customer_schema.jsonify(customer), 200
     return jsonify({"error": "Customer not found"}), 404


#UPDATE SPECIFIC CUSTOMER
@customers_bp.route("/", methods=['PUT'])
@token_required
def update_customer(customer_id):
    customer = db.session.get(Customer, customer_id)

    if not customer:
        return jsonify({"error": "Customer not found."}), 404
    
    try:
        customer_data = customer_schema.load(request.json)
    except ValidationError as e:
        return jsonify(e.messages), 400
    
    for key, value in customer_data.items():
        setattr(customer, key, value)

    db.session.commit()
    return customer_schema.jsonify(customer), 200

#DELETE SPECIFIC CUSTOMER
@customers_bp.route("/", methods=['DELETE'])
@token_required
def delete_customer(customer_id):
    customer = db.session.get(Customer, customer_id)

    if not customer:
        return jsonify({"error": "Customer not found."}), 404
    
    db.session.delete(customer)
    db.session.commit()
    return jsonify({"message": f'Customer id: {customer_id}, successfully deleted.'}), 200