# app/blueprints/mechanics/routes.py
from .schemas import mechanic_schema, mechanics_schema
from flask import request, jsonify
from marshmallow import ValidationError
from sqlalchemy import select
from app.models import Mechanic, db
from .import mechanics_bp


#Create a new mechanic

@mechanics_bp.route("/", methods=['POST'])
def create_mechanic():
     try: 
          mechanic_data = mechanic_schema.load(request.json) # Validate and deserialize input
     except ValidationError as e:
          return jsonify(e.messages),400
     
     query = select(Mechanic). where(Mechanic.email == mechanic_data['email']) #checking our db for a mechanic with this email
     existing_mechanic = db.session.execute(query).scalars().all()
     if existing_mechanic:
          return jsonify({"error": "Email already associated with another account."}), 400
     
     new_mechanic = Mechanic(**mechanic_data)
     db.session.add(new_mechanic)
     db.session.commit()
     return mechanic_schema.jsonify(new_mechanic), 201

#Get all mechanics
@mechanics_bp.route("/", methods=['GET'])
def get_mechanics():
     query = select(Mechanic)
     mechanics = db.session.execute(query).scalars().all()

     return mechanics_schema.jsonify(mechanics)

#UPDATE SPECIFIC MECHANIC
@mechanics_bp.route("/<int:mechanic_id>", methods=['PUT'])
def update_mechanic(mechanic_id):
    mechanic = db.session.get(Mechanic, mechanic_id)

    if not mechanic:
        return jsonify({"error": "Mechanic not found."}), 404
    
    try:
        mechanic_data = mechanic_schema.load(request.json)
    except ValidationError as e:
        return jsonify(e.messages), 400

    for key, value in mechanic_data.items():
        setattr(mechanic, key, value)

    db.session.commit()
    return mechanic_schema.jsonify(mechanic), 200

#DELETE SPECIFIC MECHANIC
@mechanics_bp.route("/<int:mechanic_id>", methods=['DELETE'])
def delete_mechanic(mechanic_id):
    mechanic = db.session.get(Mechanic, mechanic_id)

    if not mechanic:
        return jsonify({"error": "Mechanic not found."}), 404

    db.session.delete(mechanic)
    db.session.commit()
    return jsonify({"message": f'Mechanic id: {mechanic_id}, successfully deleted.'}), 200

#Get specific mechanic
@mechanics_bp.route("/<int:mechanic_id>", methods=["GET"])
def get_mechanic(mechanic_id):
    mechanic = db.session.get(Mechanic, mechanic_id)

    if mechanic:
          return mechanic_schema.jsonify(mechanic), 200
    return jsonify({"error": "Mechanic not found"}), 404

@mechanics_bp.route("/popular", methods=["GET"])
def popular_mechanics():
    query = select(Mechanic)
    mechanics = db.session.execute(query).scalars().all()

    mechanics.sort(key = lambda mechanic: len(mechanic.service_tickets), reverse=True)
    
    return mechanics_schema.jsonify(mechanics)

@mechanics_bp.route("/searchs", methods=["GET"])
def search_mechanics():
    name = request.args.get("name")
    
    query = select(Mechanic).where(Mechanic.name.like(f'%{name}%'))
    mechanics = db.session.execute(query).scalars().all()

    return mechanics_schema.jsonify(mechanics)